refactor(contexts): split hooks out of provider files for fast refresh
Resolves the react-doctor / eslint warnings surfaced on #31:

- only-export-components (theme, toast, release-notes contexts): move each
  context object + its use* hook (+ shared types) into a sibling
  *-context-value.ts module so the .tsx files export only their Provider
  component. Consumers import the hooks/types from the value modules.
- exhaustive-deps (toast-context): define hideToast before showToast and
  add it to showToast's dependency array.
