+++ src/contexts/release-notes-context.tsx
@@ -1,24 +1,10 @@
-import React, { createContext, useContext, useEffect, useState, useCallback, useMemo } from "react";
+import React, { useEffect, useState, useCallback, useMemo } from "react";
 import { getVersion } from "@tauri-apps/api/app";
 import { getChangelogEntries, type ChangelogEntry } from "@utils/changelog";
 import { ReleaseNotesModal } from "@components/release-notes/release-notes-modal";
+import { ReleaseNotesContext } from "./release-notes-context-value";
 
 const SEEN_VERSION_KEY = "excelerate:releaseNotesSeenVersion";
-
-interface ReleaseNotesContextType {
-  currentVersion: string;
-  openReleaseNotes: () => void;
-}
-
-const ReleaseNotesContext = createContext<ReleaseNotesContextType | undefined>(undefined);
-
-export const useReleaseNotes = () => {
-  const context = useContext(ReleaseNotesContext);
-  if (!context) {
-    throw new Error("useReleaseNotes must be used within a ReleaseNotesProvider");
-  }
-  return context;
-};
 
 interface ReleaseNotesProviderProps {
   children: React.ReactNode;
