+++ src/contexts/theme-context.tsx
@@ -1,21 +1,5 @@
-import React, { createContext, useContext, useEffect, useState, useCallback, useMemo } from "react";
-
-type Theme = "light" | "dark";
-
-interface ThemeContextType {
-  theme: Theme;
-  toggleTheme: () => void;
-}
-
-const ThemeContext = createContext<ThemeContextType | undefined>(undefined);
-
-export const useTheme = () => {
-  const context = useContext(ThemeContext);
-  if (!context) {
-    throw new Error("useTheme must be used within a ThemeProvider");
-  }
-  return context;
-};
+import React, { useEffect, useState, useCallback, useMemo } from "react";
+import { ThemeContext, type Theme } from "./theme-context-value";
 
 interface ThemeProviderProps {
   children: React.ReactNode;
