+++ src/contexts/toast-context-value.ts
@@ -1,0 +1,26 @@
+import { createContext, useContext } from "react";
+
+export type ToastType = "success" | "error" | "warning" | "info";
+
+export interface Toast {
+  id: string;
+  title: string;
+  description?: string;
+  type: ToastType;
+  duration?: number;
+}
+
+export interface ToastContextType {
+  showToast: (toast: Omit<Toast, "id">) => void;
+  hideToast: (id: string) => void;
+}
+
+export const ToastContext = createContext<ToastContextType | undefined>(undefined);
+
+export const useToast = () => {
+  const context = useContext(ToastContext);
+  if (!context) {
+    throw new Error("useToast must be used within a ToastProvider");
+  }
+  return context;
+};
